LNA working, removed template

+++ TLS/HTTP/BlockCypher.h
@@ -49,7 +49,7 @@ protected:
 
 	static void xor(uint8_t* target, 
                     const uint8_t* src, 
-                    int len);
+                    size_t len);
 
 	size_t mBlockSize;
 	size_t mKeySize;
@@ -60,7 +60,7 @@ private:
                             uint8_t* dest,
 	                  const uint8_t* iv,
                       const uint8_t* key,
-                      const uint32_t blockCount,
+                      const size_t   blockCount,
                       const Mode	 mode,
                       const opType   operation);
 };
