+++ TLS/HTTP/HttpUtil.cpp
@@ -9,7 +9,7 @@
 static const std::string base64_chars = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
 										"abcdefghijklmnopqrstuvwxyz"
 										"0123456789+/";
-std::string base64_encode(uint8_t const* bytes_to_encode, unsigned int in_len) {
+std::string base64_encode(uint8_t const* bytes_to_encode, size_t in_len) {
     
     std::string ret;
     int i = 0;
@@ -58,7 +58,7 @@ static inline bool is_base64(uint8_t c) {
 }
 
 std::string base64_decode(std::string const& encoded_string) {
-    int in_len = encoded_string.size();
+    size_t in_len = encoded_string.size();
     int i = 0;
     int j = 0;
     int in_ = 0;
@@ -69,7 +69,7 @@ std::string base64_decode(std::string const& encoded_string) {
         char_array_4[i++] = encoded_string[in_]; in_++;
         if (i == 4) {
             for (i = 0; i <4; i++)
-                char_array_4[i] = base64_chars.find(char_array_4[i]);
+                char_array_4[i] = (uint8_t)base64_chars.find(char_array_4[i]);
 
             char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
             char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -86,7 +86,7 @@ std::string base64_decode(std::string const& encoded_string) {
             char_array_4[j] = 0;
 
         for (j = 0; j <4; j++)
-            char_array_4[j] = base64_chars.find(char_array_4[j]);
+            char_array_4[j] = (uint8_t)base64_chars.find(char_array_4[j]);
 
         char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
         char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -99,7 +99,7 @@ std::string base64_decode(std::string const& encoded_string) {
 }
 
 
-std::string base16_encode(uint8_t const* bytesToEncode, uint32_t byteLength)
+std::string base16_encode(uint8_t const* bytesToEncode, size_t byteLength)
 {
 	static const std::string base16_chars = "0123456789ABCDEF";
 	std::string encoding("0x");
